requirements.txt
blinker==1.7.0
click==8.1.7
colorama==0.4.6
emoji==2.11.0
Flask==3.0.2
Flask-Mail==0.9.1
gunicorn==21.2.0
itsdangerous==2.1.2
Jinja2==3.1.3
MarkupSafe==2.1.5
mysql-connector==2.2.9
packaging==24.0
Werkzeug==3.0.1
